dont pass arbasicobjkinds around
