+++ tools/clang/include/clang/Sema/SemaHLSL.h
@@ -24,208 +24,6 @@
 #include "clang/Sema/Template.h"
 #include "clang/Sema/TemplateDeduction.h"
 
-enum ArBasicKind {
-  AR_BASIC_BOOL,
-  AR_BASIC_LITERAL_FLOAT,
-  AR_BASIC_FLOAT16,
-  AR_BASIC_FLOAT32_PARTIAL_PRECISION,
-  AR_BASIC_FLOAT32,
-  AR_BASIC_FLOAT64,
-  AR_BASIC_LITERAL_INT,
-  AR_BASIC_INT8,
-  AR_BASIC_UINT8,
-  AR_BASIC_INT16,
-  AR_BASIC_UINT16,
-  AR_BASIC_INT32,
-  AR_BASIC_UINT32,
-  AR_BASIC_INT64,
-  AR_BASIC_UINT64,
-
-  AR_BASIC_MIN10FLOAT,
-  AR_BASIC_MIN16FLOAT,
-  AR_BASIC_MIN12INT,
-  AR_BASIC_MIN16INT,
-  AR_BASIC_MIN16UINT,
-  AR_BASIC_INT8_4PACKED,
-  AR_BASIC_UINT8_4PACKED,
-  AR_BASIC_ENUM,
-
-  AR_BASIC_COUNT,
-
-  //
-  // Pseudo-entries for intrinsic tables and such.
-  //
-
-  AR_BASIC_NONE,
-  AR_BASIC_UNKNOWN,
-  AR_BASIC_NOCAST,
-  AR_BASIC_DEPENDENT,
-  //
-  // The following pseudo-entries represent higher-level
-  // object types that are treated as units.
-  //
-
-  AR_BASIC_POINTER,
-  AR_BASIC_ENUM_CLASS,
-
-  AR_OBJECT_NULL,
-  AR_OBJECT_STRING_LITERAL,
-  AR_OBJECT_STRING,
-
-  // AR_OBJECT_TEXTURE,
-  AR_OBJECT_TEXTURE1D,
-  AR_OBJECT_TEXTURE1D_ARRAY,
-  AR_OBJECT_TEXTURE2D,
-  AR_OBJECT_TEXTURE2D_ARRAY,
-  AR_OBJECT_TEXTURE3D,
-  AR_OBJECT_TEXTURECUBE,
-  AR_OBJECT_TEXTURECUBE_ARRAY,
-  AR_OBJECT_TEXTURE2DMS,
-  AR_OBJECT_TEXTURE2DMS_ARRAY,
-
-  AR_OBJECT_SAMPLER,
-  AR_OBJECT_SAMPLER1D,
-  AR_OBJECT_SAMPLER2D,
-  AR_OBJECT_SAMPLER3D,
-  AR_OBJECT_SAMPLERCUBE,
-  AR_OBJECT_SAMPLERCOMPARISON,
-
-  AR_OBJECT_BUFFER,
-
-  //
-  // View objects are only used as variable/types within the Effects
-  // framework, for example in calls to OMSetRenderTargets.
-  //
-
-  AR_OBJECT_RENDERTARGETVIEW,
-  AR_OBJECT_DEPTHSTENCILVIEW,
-
-  //
-  // Shader objects are only used as variable/types within the Effects
-  // framework, for example as a result of CompileShader().
-  //
-
-  AR_OBJECT_COMPUTESHADER,
-  AR_OBJECT_DOMAINSHADER,
-  AR_OBJECT_GEOMETRYSHADER,
-  AR_OBJECT_HULLSHADER,
-  AR_OBJECT_PIXELSHADER,
-  AR_OBJECT_VERTEXSHADER,
-  AR_OBJECT_PIXELFRAGMENT,
-  AR_OBJECT_VERTEXFRAGMENT,
-
-  AR_OBJECT_STATEBLOCK,
-
-  AR_OBJECT_RASTERIZER,
-  AR_OBJECT_DEPTHSTENCIL,
-  AR_OBJECT_BLEND,
-
-  AR_OBJECT_POINTSTREAM,
-  AR_OBJECT_LINESTREAM,
-  AR_OBJECT_TRIANGLESTREAM,
-
-  AR_OBJECT_INPUTPATCH,
-  AR_OBJECT_OUTPUTPATCH,
-
-  AR_OBJECT_RWTEXTURE1D,
-  AR_OBJECT_RWTEXTURE1D_ARRAY,
-  AR_OBJECT_RWTEXTURE2D,
-  AR_OBJECT_RWTEXTURE2D_ARRAY,
-  AR_OBJECT_RWTEXTURE3D,
-  AR_OBJECT_RWBUFFER,
-
-  AR_OBJECT_BYTEADDRESS_BUFFER,
-  AR_OBJECT_RWBYTEADDRESS_BUFFER,
-  AR_OBJECT_STRUCTURED_BUFFER,
-  AR_OBJECT_RWSTRUCTURED_BUFFER,
-  AR_OBJECT_RWSTRUCTURED_BUFFER_ALLOC,
-  AR_OBJECT_RWSTRUCTURED_BUFFER_CONSUME,
-  AR_OBJECT_APPEND_STRUCTURED_BUFFER,
-  AR_OBJECT_CONSUME_STRUCTURED_BUFFER,
-
-  AR_OBJECT_CONSTANT_BUFFER,
-  AR_OBJECT_TEXTURE_BUFFER,
-
-  AR_OBJECT_ROVBUFFER,
-  AR_OBJECT_ROVBYTEADDRESS_BUFFER,
-  AR_OBJECT_ROVSTRUCTURED_BUFFER,
-  AR_OBJECT_ROVTEXTURE1D,
-  AR_OBJECT_ROVTEXTURE1D_ARRAY,
-  AR_OBJECT_ROVTEXTURE2D,
-  AR_OBJECT_ROVTEXTURE2D_ARRAY,
-  AR_OBJECT_ROVTEXTURE3D,
-
-  AR_OBJECT_FEEDBACKTEXTURE2D,
-  AR_OBJECT_FEEDBACKTEXTURE2D_ARRAY,
-
-// SPIRV change starts
-#ifdef ENABLE_SPIRV_CODEGEN
-  AR_OBJECT_VK_SUBPASS_INPUT,
-  AR_OBJECT_VK_SUBPASS_INPUT_MS,
-  AR_OBJECT_VK_SPIRV_TYPE,
-  AR_OBJECT_VK_SPIRV_OPAQUE_TYPE,
-  AR_OBJECT_VK_INTEGRAL_CONSTANT,
-  AR_OBJECT_VK_LITERAL,
-  AR_OBJECT_VK_SPV_INTRINSIC_TYPE,
-  AR_OBJECT_VK_SPV_INTRINSIC_RESULT_ID,
-#endif // ENABLE_SPIRV_CODEGEN
-  // SPIRV change ends
-
-  AR_OBJECT_INNER, // Used for internal type object
-
-  AR_OBJECT_LEGACY_EFFECT,
-
-  AR_OBJECT_WAVE,
-
-  AR_OBJECT_RAY_DESC,
-  AR_OBJECT_ACCELERATION_STRUCT,
-  AR_OBJECT_USER_DEFINED_TYPE,
-  AR_OBJECT_TRIANGLE_INTERSECTION_ATTRIBUTES,
-
-  // subobjects
-  AR_OBJECT_STATE_OBJECT_CONFIG,
-  AR_OBJECT_GLOBAL_ROOT_SIGNATURE,
-  AR_OBJECT_LOCAL_ROOT_SIGNATURE,
-  AR_OBJECT_SUBOBJECT_TO_EXPORTS_ASSOC,
-  AR_OBJECT_RAYTRACING_SHADER_CONFIG,
-  AR_OBJECT_RAYTRACING_PIPELINE_CONFIG,
-  AR_OBJECT_TRIANGLE_HIT_GROUP,
-  AR_OBJECT_PROCEDURAL_PRIMITIVE_HIT_GROUP,
-  AR_OBJECT_RAYTRACING_PIPELINE_CONFIG1,
-
-  // RayQuery
-  AR_OBJECT_RAY_QUERY,
-
-  // Heap Resource
-  AR_OBJECT_HEAP_RESOURCE,
-  AR_OBJECT_HEAP_SAMPLER,
-
-  AR_OBJECT_RWTEXTURE2DMS,
-  AR_OBJECT_RWTEXTURE2DMS_ARRAY,
-
-  // Work Graphs
-  AR_OBJECT_EMPTY_NODE_INPUT,
-  AR_OBJECT_DISPATCH_NODE_INPUT_RECORD,
-  AR_OBJECT_RWDISPATCH_NODE_INPUT_RECORD,
-  AR_OBJECT_GROUP_NODE_INPUT_RECORDS,
-  AR_OBJECT_RWGROUP_NODE_INPUT_RECORDS,
-  AR_OBJECT_THREAD_NODE_INPUT_RECORD,
-  AR_OBJECT_RWTHREAD_NODE_INPUT_RECORD,
-
-  AR_OBJECT_NODE_OUTPUT,
-  AR_OBJECT_EMPTY_NODE_OUTPUT,
-  AR_OBJECT_NODE_OUTPUT_ARRAY,
-  AR_OBJECT_EMPTY_NODE_OUTPUT_ARRAY,
-
-  AR_OBJECT_THREAD_NODE_OUTPUT_RECORDS,
-  AR_OBJECT_GROUP_NODE_OUTPUT_RECORDS,
-
-  // Shader Execution Reordering
-  AR_OBJECT_HIT_OBJECT,
-
-  AR_BASIC_MAXIMUM_COUNT
-};
-
 // Forward declarations.
 struct IDxcIntrinsicTable;
 namespace clang {
